+++ lib/automated_testing_framework.dart
@@ -1,10 +1,4 @@
-export 'src/driver/model/drivable_device.dart';
-export 'src/driver/model/driven_test_status.dart';
-export 'src/driver/model/driver_test_request.dart';
-export 'src/driver/model/external_test_driver.dart';
-export 'src/driver/model/test_device_status.dart';
-
-export 'src/driver/utils/driver_signature_helper.dart';
+export 'package:automated_testing_framework_models/automated_testing_framework_models.dart';
 
 export 'src/driver/widgets/test_driver_status.dart';
 
@@ -28,10 +22,6 @@ export 'src/framework/forms/tap_form.dart';
 export 'src/framework/forms/test_step_form.dart';
 
 export 'src/framework/models/available_test_step.dart';
-export 'src/framework/models/test_app_settings.dart';
-export 'src/framework/models/test_step.dart';
-export 'src/framework/models/testable_gesture_action.dart';
-export 'src/framework/models/testable_gestures.dart';
 
 export 'src/framework/pages/available_tests_page.dart';
 export 'src/framework/pages/test_steps_page.dart';
@@ -43,13 +33,7 @@ export 'src/framework/widgets/test_step_picker.dart';
 export 'src/registry/test_step_builder.dart';
 export 'src/registry/test_step_registry.dart';
 
-export 'src/report/models/golden_test_images.dart';
 export 'src/report/models/test_device_info.dart';
-export 'src/report/models/test_image.dart';
-export 'src/report/models/test_report.dart';
-export 'src/report/models/test_report_step.dart';
-export 'src/report/models/test_suite_report.dart';
-export 'src/report/models/test_suite_result.dart';
 
 export 'src/report/pages/test_report_page.dart';
 export 'src/report/pages/test_suite_report_page.dart';
@@ -61,9 +45,7 @@ export 'src/runner/components/test_controller.dart';
 
 export 'src/runner/models/capture_context.dart';
 export 'src/runner/models/progress_value.dart';
-export 'src/runner/models/test.dart';
 export 'src/runner/models/test_runner_theme_data.dart';
-export 'src/runner/models/test_step_delays.dart';
 
 export 'src/runner/steps/assert_error_step.dart';
 export 'src/runner/steps/assert_value_step.dart';
@@ -88,12 +70,6 @@ export 'src/runner/widgets/test_runner.dart';
 export 'src/store/components/asset_test_store.dart';
 export 'src/store/components/clipboard_test_store.dart';
 export 'src/store/components/test_store.dart';
-
-export 'src/store/loaders/async_test_loader.dart';
-export 'src/store/loaders/memory_test_loader.dart';
-
-export 'src/store/models/pending_test.dart';
-export 'src/store/models/test_loader.dart';
 
 export 'src/store/typedefs/golden_image_writer.dart';
 export 'src/store/typedefs/test_image_reader.dart';