+++ lib/src/driver/model/test_device_status.dart
@@ -1,8 +1,0 @@
-class TestDeviceStatus {
-  TestDeviceStatus._();
-
-  static const available = 'available';
-  static const connected = 'connected';
-  static const offline = 'offline';
-  static const running = 'running';
-}