+++ lib/src/report/models/test_device_info.dart
@@ -4,107 +4,45 @@ import 'package:automated_testing_framework/automated_testing_framework.dart';
 import 'package:device_info/device_info.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:json_class/json_class.dart';
 import 'package:package_info/package_info.dart';
 import 'package:websafe_platform/websafe_platform.dart';
 
 /// Container class for the device information for the device the test is being
 /// executed on.
 @immutable
-class TestDeviceInfo extends JsonClass {
-  TestDeviceInfo.custom({
-    @required this.appIdentifier,
-    @required this.brand,
-    @required this.buildNumber,
-    @required this.device,
-    @required this.devicePixelRatio,
-    @required this.dips,
-    @required this.id,
-    @required this.manufacturer,
-    @required this.model,
-    @required this.os,
-    @required this.orientation,
-    bool physicalDevice = true,
-    @required this.pixels,
-    @required this.systemVersion,
-  }) : physicalDevice = physicalDevice ?? true;
+class TestDeviceInfo extends BaseTestDeviceInfo {
+  TestDeviceInfo._internal(BaseTestDeviceInfo info)
+      : super.custom(
+          appIdentifier: info.appIdentifier,
+          brand: info.brand,
+          buildNumber: info.buildNumber,
+          device: info.device,
+          devicePixelRatio: info.devicePixelRatio,
+          dips: info.dips,
+          id: info.id,
+          manufacturer: info.manufacturer,
+          model: info.model,
+          orientation: info.orientation,
+          os: info.os,
+          physicalDevice: info.physicalDevice,
+          pixels: info.pixels,
+          systemVersion: info.systemVersion,
+        );
 
   static Completer<TestDeviceInfo> _completer;
   static TestDeviceInfo _instance;
 
   static TestDeviceInfo get instance => _instance ?? initialize(null);
 
-  final String appIdentifier;
-  final String brand;
-  final String buildNumber;
-  final String device;
-  final double devicePixelRatio;
-  final Size dips;
-  final String id;
-  final String manufacturer;
-  final String model;
-  final String orientation;
-  final String os;
-  final bool physicalDevice;
-  final Size pixels;
-  final String systemVersion;
-
   static TestDeviceInfo fromDynamic(dynamic map) {
     TestDeviceInfo result;
 
     if (map != null) {
-      var screen = map['screen'] ?? {};
-      var dips = screen['dips'];
-      var pixels = screen['pixels'];
-
-      result = TestDeviceInfo.custom(
-        appIdentifier: map['appIdentifier'],
-        brand: map['brand'],
-        buildNumber: map['buildNumber'],
-        device: map['device'],
-        devicePixelRatio: JsonClass.parseDouble(screen['devicePixelRatio']),
-        dips: dips == null
-            ? null
-            : Size(
-                JsonClass.parseDouble(
-                  dips['width'],
-                ),
-                JsonClass.parseDouble(
-                  dips['height'],
-                ),
-              ),
-        id: map['id'],
-        manufacturer: map['manufacturer'],
-        model: map['model'],
-        orientation: map['orientation'],
-        os: map['os'],
-        physicalDevice: JsonClass.parseBool(map['physicalDevice']),
-        pixels: pixels == null
-            ? null
-            : Size(
-                JsonClass.parseDouble(
-                  pixels['width'],
-                ),
-                JsonClass.parseDouble(
-                  pixels['height'],
-                ),
-              ),
-        systemVersion: map['systemVersion'],
-      );
+      result = TestDeviceInfo._internal(BaseTestDeviceInfo.fromDynamic(map));
     }
 
     return result;
   }
-
-  String get deviceSignature => [
-        appIdentifier,
-        buildNumber,
-        os,
-        manufacturer,
-        model,
-        physicalDevice,
-        systemVersion,
-      ].join('_');
 
   static Future<TestDeviceInfo> initialize(BuildContext context) async {
     var result = _instance;
@@ -122,14 +60,14 @@ class TestDeviceInfo extends JsonClass {
       String buildNumber;
       String device;
       double devicePixelRatio;
-      Size dips;
+      BaseSize dips;
       var id = '<unknown>';
       String manufacturer;
       String model;
       String os;
       String orientation;
       bool physicalDevice;
-      Size pixels;
+      BaseSize pixels;
       String systemVersion;
       var wsp = WebsafePlatform();
 
@@ -214,13 +152,13 @@ class TestDeviceInfo extends JsonClass {
         try {
           var mq = MediaQuery.of(context);
 
-          dips = Size(
+          dips = BaseSize(
             mq.size.width,
             mq.size.height,
           );
           orientation = dips.width >= dips.height ? 'landscape' : 'portrait';
           devicePixelRatio = mq.devicePixelRatio;
-          pixels = Size(
+          pixels = BaseSize(
             mq.size.width * devicePixelRatio,
             mq.size.height * devicePixelRatio,
           );
@@ -228,21 +166,24 @@ class TestDeviceInfo extends JsonClass {
           // no-op, we don't know the screen, but let's not make a big fuss.
         }
       }
-      result = TestDeviceInfo.custom(
-        appIdentifier: TestAppSettings.settings.appIdentifier ?? appIdentifier,
-        brand: brand,
-        buildNumber: buildNumber,
-        device: device,
-        devicePixelRatio: devicePixelRatio,
-        dips: dips,
-        id: id,
-        manufacturer: manufacturer,
-        model: model,
-        orientation: orientation,
-        os: os,
-        physicalDevice: physicalDevice,
-        pixels: pixels,
-        systemVersion: systemVersion,
+      result = TestDeviceInfo._internal(
+        BaseTestDeviceInfo.custom(
+          appIdentifier:
+              TestAppSettings.settings.appIdentifier ?? appIdentifier,
+          brand: brand,
+          buildNumber: buildNumber,
+          device: device,
+          devicePixelRatio: devicePixelRatio,
+          dips: dips,
+          id: id,
+          manufacturer: manufacturer,
+          model: model,
+          orientation: orientation,
+          os: os,
+          physicalDevice: physicalDevice,
+          pixels: pixels,
+          systemVersion: systemVersion,
+        ),
       );
 
       _instance = result;
@@ -252,36 +193,4 @@ class TestDeviceInfo extends JsonClass {
 
     return result;
   }
-
-  @override
-  Map<String, dynamic> toJson() => {
-        'appIdentifier': appIdentifier,
-        'brand': brand,
-        'buildNumber': buildNumber,
-        'device': device,
-        'id': id,
-        'manufacturer': manufacturer,
-        'model': model,
-        'orientation': orientation,
-        'os': os,
-        'physicalDevice': physicalDevice,
-        'screen': devicePixelRatio == null
-            ? null
-            : {
-                'devicePixelRatio': devicePixelRatio,
-                'dips': dips == null
-                    ? null
-                    : {
-                        'height': dips.height,
-                        'width': dips.width,
-                      },
-                'pixels': pixels == null
-                    ? null
-                    : {
-                        'height': pixels.height,
-                        'width': pixels.width,
-                      },
-              },
-        'systemVersion': systemVersion,
-      };
 }