+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.2.18+2] - November 3rd, 2020
+
+* Added timestamp to `TestSuiteReportPage`.
+
+
 ## [1.2.18+1] - October 28th, 2020
 
 * Added `appIdentifier` to `ExternalTestDriver` and `GoldenTestImages` id.